+++ src/main/java/it/unipi/hadoop/TotalCharacterCount.java
@@ -1,7 +1,7 @@
 package it.unipi.hadoop;
 
 import org.apache.commons.lang.StringUtils;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 import org.apache.hadoop.mapreduce.Reducer;
@@ -12,14 +12,14 @@ import java.util.Map;
 
 public class TotalCharacterCount {
 
-    public static class TotalCountMapper extends Mapper<Object, Text, Text, IntWritable> {
+    public static class TotalCountMapper extends Mapper<Object, Text, Text, LongWritable> {
 
         private final Text outputKey = new Text("total");
-        private final IntWritable outputValue = new IntWritable();
-        private Map<Character, Integer> totalCount; // In-mapper combining
+        private final LongWritable outputValue = new LongWritable();
+        private Map<Character, Long> totalCount; // In-mapper combining
 
         @Override
-        protected void setup(Mapper<Object, Text, Text, IntWritable>.Context context) throws IOException, InterruptedException {
+        protected void setup(Mapper<Object, Text, Text, LongWritable>.Context context) throws IOException, InterruptedException {
             totalCount = new HashMap<>();
         }
 
@@ -30,30 +30,29 @@ public class TotalCharacterCount {
 
             for (char c : tokenizer) {
                 if (((int) c >= 97 && (int) c <= 122)) {
-                    totalCount.merge('T', 1, Integer::sum);
+                    totalCount.merge('T', 1L, Long::sum);
                 }
             }
         }
 
         @Override
-        protected void cleanup(Mapper<Object, Text, Text, IntWritable>.Context context) throws IOException, InterruptedException {
-            for (Map.Entry<Character, Integer> entry : totalCount.entrySet()) {
+        protected void cleanup(Mapper<Object, Text, Text, LongWritable>.Context context) throws IOException, InterruptedException {
+            for (Map.Entry<Character, Long> entry : totalCount.entrySet()) {
                 outputValue.set(entry.getValue());
                 context.write(outputKey, outputValue);
             }
         }
     }
 
-    public static class TotalCountReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+    public static class TotalCountReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
 
         @Override
-        protected void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
-            int sum = 0;
-            for (IntWritable val : values) {
+        protected void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
+            long sum = 0L;
+            for (LongWritable val : values) {
                 sum += val.get();
             }
-            context.write(key, new IntWritable(sum));
-
+            context.write(key, new LongWritable(sum));
         }
     }
 }
